bug: fixing eval and tidying up
