+++ evaluation/run_evaluation.py
@@ -3,8 +3,10 @@ import os
 from src.collection.evaluate_collection import (
     assess_retrieval_accuracy,
     get_data_for_evaluation,
-    load_qdrant_client,
+    get_all_regex_ids,
+    assess_scroll_retrieval,
 )
+from src.utils.utils import load_qdrant_client
 
 # Get env vars
 PUBLISHING_PROJECT_ID = os.getenv("PUBLISHING_PROJECT_ID")
@@ -24,13 +26,26 @@ def main():
         evaluation_table=EVALUATION_TABLE,
     )
 
+    regex_ids = get_all_regex_ids(data)
+
     # Assess the retrieval accuracy
-    assess_retrieval_accuracy(
+    ss_results = assess_retrieval_accuracy(
         client=client,
         collection_name=COLLECTION_NAME,
-        labels=data,
+        data=data,
+        regex_ids=regex_ids,
         k_threshold=100,
     )
+    print(f"Dot product search n results: {len(ss_results)}")
+
+    # Assess the scroll retrieval accuracy
+    scroll_results = assess_scroll_retrieval(
+        client=client,
+        collection_name=COLLECTION_NAME,
+        data=data,
+        regex_ids=regex_ids,
+    )
+    print(f"Scroll n results: {len(scroll_results)}")
 
 
 if __name__ == "__main__":
